Fix bug where ParserGenerator was not actually preferring root discriminators for prefixes.
Additionally, investigate issue of grammars which are not differentiable under the basic lookahead rules and
actually require proactive discriminator symbol generation to avoid infinite
recursion. For now, we have new test cases in place documenting this behavior
and a hard NotSupportedException throw for when we beleive we've hit this case.

Solving this problem likely requires a substantial refactor of the core algorithm flow
to allow for a more speculative approach where we consider multiple paths to parsing
a given NodeContext and choose the shortest one.